backend/requirements.txt
fastapi
uvicorn
pydantic
PyJWT
python-multipart